fix(globalization): read/write pvar floats as InvariantCulture

+++ Assets/Forge/Scripts/Collision/Editor/CollisionBaker.cs
@@ -151,7 +151,7 @@ public static class CollisionBaker
                     var go = new GameObject(reason);
                     go.transform.SetParent(collisionResultsVisualizer.transform, false);
                     var node = go.AddComponent<CollisionResultsVisualizerNode>();
-                    node.transform.position = new Vector3(float.Parse(x) + 0.5f, float.Parse(y) + 0.5f, float.Parse(z) + 0.5f) * 4;
+                    node.transform.position = new Vector3(int.Parse(x) + 0.5f, int.Parse(y) + 0.5f, int.Parse(z) + 0.5f) * 4;
                     //Debug.Log($"{reason}: {node.transform.position}");
                 }
